requirements.txt
langchain
langchain-openai
langgraph
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
streamlit